requirements.txt
pandas==1.2.1
pyyaml==5.4.1
nested_lookup==0.2.21